패턴 클래스 분리
