+++ connector_predict.py
@@ -3,22 +3,6 @@ import os
 import predict
 
 
-class Singleton(type):
-    """
-    Define an Instance operation that lets clients access its unique
-    instance.
-    """
-
-    def __init__(cls, name, bases, attrs, **kwargs):
-        super().__init__(name, bases, attrs)
-        cls._instance = None
-
-    def __call__(cls, *args, **kwargs):
-        if cls._instance is None:
-            cls._instance = super().__call__(*args, **kwargs)
-        return cls._instance
-
-
 class Connector():
     def __init__(self, image_file):
         MODEL_DIR = 'imagenet_model'
