+++ predict.py
@@ -8,28 +8,16 @@ import os.path
 import re
 import sys
 import tarfile
-
 import numpy as np
 import tensorflow as tf
 from six.moves import urllib
+from pattern import Singleton
 
-
-class Singleton(type):
-    """
-    Define an Instance operation that lets clients access its unique
-    instance.
-    """
-
-    def __init__(cls, name, bases, attrs, **kwargs):
-        super().__init__(name, bases, attrs)
-        cls._instance = None
-
-    def __call__(cls, *args, **kwargs):
-        if cls._instance is None:
-            cls._instance = super().__call__(*args, **kwargs)
-        return cls._instance
-
-
+"""
+Predict
+ref
+http://solarisailab.com/archives/1422
+"""
 class Predict(metaclass=Singleton):
     def __init__(self, model_dir, num_top_predictions, DATA_URL, pbtxt, uid):
         self.model_dir = model_dir
